requirements.txt
discord.py~=1.7.3
spotipy~=2.18.0
pafy~=0.5.5
python-dotenv~=0.18.0
ytmusicapi~=0.14.3
aiohttp~=3.7.4.post0
pip~=21.2.1
py~=1.10.0
setuptools~=54.1.1
youtube-dl~=2021.6.6
nest_asyncio~=1.5.1
PyNaCl~=1.4.0